+++ tests/testthat/test-tsc_chunk_post.R
@@ -1,0 +1,5 @@
+test_that("tsc_chunk_post works", {
+  skip("TODO")
+})
+
+# use_r("tsc_chunk_post")  # nolint