+++ R/tsc_occ_obs_post.R
@@ -13,11 +13,12 @@
 #' testthat::skip_if_not(tsc_works())
 #' data("phys_sample")
 #' x <- tsc_occ_obs_post(phys_sample,
-#'                       obstype="PhysicalSample",
-#'                       chunksize = 10,
-#'                       api_url = get_tsc_test_api_url(),
-#'                       api_token = get_tsc_test_api_token(),
-#'                       verbose = TRUE)))
+#'   obstype = "PhysicalSample",
+#'   chunksize = 10,
+#'   api_url = get_tsc_test_api_url(),
+#'   api_token = get_tsc_test_api_token(),
+#'   verbose = TRUE
+#' )
 #' }
 tsc_occ_obs_post <- function(data,
                              obstype = "PhysicalSample",