+++ tests/testthat/test-dl_mdbzip.R
@@ -1,0 +1,5 @@
+test_that("dl_mdbzip works", {
+  skip("TODO")
+})
+
+# use_r("dl_mdbzip")  # nolint