+++ tests/testthat/test-tsc_occ_obs_post.R
@@ -6,11 +6,12 @@ test_that("tsc_occ_obs_post can upload valid example data", {
   testthat::expect_message(
     suppressWarnings(
       x <- tsc_occ_obs_post(phys_sample,
-                            obstype="PhysicalSample",
-                            chunksize = 10,
-                            api_url = get_tsc_test_api_url(),
-                            api_token = get_tsc_test_api_token(),
-                            verbose = TRUE)
+        obstype = "PhysicalSample",
+        chunksize = 10,
+        api_url = get_tsc_test_api_url(),
+        api_token = get_tsc_test_api_token(),
+        verbose = TRUE
+      )
     )
   )