+++ tests/testthat/test-upload_file_to_ckan.R
@@ -1,0 +1,5 @@
+test_that("upload_file_to_ckan works", {
+  skip("TODO")
+})
+
+# use_r("upload_file_to_ckan")  # nolint